requirements.txt
requests==2.22.0
python-decouple==3.4
